Kyle's Edits
Added a new test case for the testing class. This test passes because the Dynamic programming solution will choose two 4's out of the coin system {1, 4, 5} to make change for 8. However, this test would fail if a greedy algorithm was used because it would choose a 5 first and then three 1's.